requirements.txt
# Core dependencies
openai>=0.27.10
llama-index>=0.6.0
prefect>=2.10.0
pydantic>=2.11.7
python-dotenv>=1.1.1
pip>=25.2
langchain
langsmith
